Cleaned the code and made a function
I removed unnecessary crap and made the whole program into a single function with a return value. Before I had a "start_phrase" and "end_phrase" which would be said around the output. This was all also loose code not contained within a function. Before it would just print out the result of the function, now it will return it as a string.